+++ Lib/Services/FoodService.cs
@@ -15,7 +15,6 @@ namespace Lib.Services
         public FoodService(ApplicationDbContext dbContext, IFoodRepository FoodRepository) {
             this.FoodRepository = FoodRepository;
             this.dbContext = dbContext;
-            
         }
         
         public void Save() {
@@ -122,9 +121,9 @@ namespace Lib.Services
 
         }
 
-        public float GetPrice ( int id)
+        public double GetPrice ( int id)
         {
-            return dbContext.Food.FirstOrDefault(s=>s.ID_Food == id).Price;
+            return dbContext.Food.FirstOrDefault(s => s.ID_Food == id).Price;
         }
     }
 }
